Moved code to common and implemented code for paging.
